playing with file parse
