+++ src/parserMDReact/parse.js
@@ -1,17 +1,17 @@
-import stateInit from '../domain/state';
-import configureReplacer from '../domain/replacer-class/configuration-react'
+// import stateInit from '../domain/state';
+// import configureReplacer from '../domain/replacer-class/configuration-react'
 
-// @todo update this method. I'm sure it can be improved.
-function parseMDReact(source, isFull) {
+// // @todo update this method. I'm sure it can be improved.
+// function parseMDReact(source, isFull) {
   
-  const state = stateInit(source);
-  configureReplacer(state, isFull);
+//   const state = stateInit(source);
+//   configureReplacer(state, isFull);
 
-  return state;
-}
+//   return state;
+// }
 
-function parseMDReactFullThing(source) {
-  return parseMDReact(source, true);
-}
+// function parseMDReactFullThing(source) {
+//   return parseMDReact(source, true);
+// }
 
-export { parseMDReact, parseMDReactFullThing };
+// export { parseMDReact, parseMDReactFullThing };
